frontend/requirements.txt
uvicorn
streamlit
requests
loguru
streamlit-drawable-canvas
pillow
numpy
fastapi
pydantic